refactor: auth 패키지 의존성 제거
