+++ src/main/java/auth/MemberDetailsService.java
@@ -1,29 +1,6 @@
 package auth;
 
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.stereotype.Service;
+public interface MemberDetailsService {
 
-@Service
-public class MemberDetailsService {
-
-    private final JdbcTemplate jdbcTemplate;
-
-    public MemberDetailsService(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
-
-    private final RowMapper<MemberDetails> rowMapper = (resultSet, rowNum) -> new MemberDetails(
-            resultSet.getLong("id"),
-            resultSet.getString("username"),
-            resultSet.getString("password"),
-            resultSet.getString("name"),
-            resultSet.getString("phone"),
-            resultSet.getString("role")
-    );
-
-    public MemberDetails loadMemberDetailsByUsername(String username) {
-        String sql = "SELECT id, username, password, name, phone, role from member where username = ?;";
-        return jdbcTemplate.queryForObject(sql, rowMapper, username);
-    }
+    MemberDetails loadMemberDetailsByUsername(String username);
 }
